improve: Option Component to ts
