+++ packages/fuselage/src/components/Options/Option/Option.spec.tsx
@@ -1,8 +1,12 @@
 import { render } from '@testing-library/react';
 import React from 'react';
 
-import { Option } from '.';
+import Option from '.';
 
 it('renders without crashing', () => {
-  render(<Option />);
+  render(
+    <Option>
+      <Option.Content>Lorem Ipsum Lorem</Option.Content>
+    </Option>
+  );
 });
