+++ packages/fuselage/src/components/Options/Option/Option.stories.mdx
@@ -1,4 +1,5 @@
-import { Meta, Canvas, ArgsTable, Story } from '@storybook/addon-docs/blocks';
+import { action } from '@storybook/addon-actions';
+import { Meta, Canvas, ArgsTable, Story } from '@storybook/addon-docs';
 import LinkTo from '@storybook/addon-links/react';
 
 import { Option } from '.';
@@ -14,26 +15,33 @@ The generic <LinkTo kind='Misc/Options/Option' story='Default'>`Option`</LinkTo>
 <Canvas>
   <Story name='Default'>
     <Box position='relative' maxWidth={250}>
-      <Option>
+      <Option onClick={action('click')}>
         <Option.Content>Lorem Ipsum Lorem</Option.Content>
       </Option>
       <Option>
-        <Option.Content>Lorem Ipsum Lorem Lorem Ipsum Lorem Lorem Ipsum Lorem Lorem Ipsum Lorem</Option.Content>
+        <Option.Content>
+          Lorem Ipsum Lorem Lorem Ipsum Lorem Lorem Ipsum Lorem Lorem Ipsum
+          Lorem
+        </Option.Content>
       </Option>
       <Option>
-        <Option.Content>Lorem Ipsum Lorem <Option.Description>Lorem Lorem Ipsum Lorem Lorem Ipsum Lorem</Option.Description></Option.Content>
+        <Option.Content>
+          Lorem Ipsum Lorem{' '}
+          <Option.Description>
+            Lorem Lorem Ipsum Lorem Lorem Ipsum Lorem
+          </Option.Description>
+        </Option.Content>
       </Option>
     </Box>
   </Story>
 </Canvas>
-
 
 ## With avatar
 
 <Canvas>
   <Story name='withAvatar'>
     <Box position='relative' maxWidth={250}>
-      <Option>
+      <Option onClick={action('click')}>
         <Option.Avatar>
           <Avatar url={exampleAvatar} size='x28' />
         </Option.Avatar>
@@ -42,14 +50,22 @@ The generic <LinkTo kind='Misc/Options/Option' story='Default'>`Option`</LinkTo>
       <Option>
         <Option.Avatar>
           <Avatar url={exampleAvatar} size='x28' />
-        </Option.Avatar>        
-        <Option.Content>Lorem Ipsum Lorem Lorem Ipsum Lorem Lorem Ipsum Lorem Lorem Ipsum Lorem</Option.Content>
+        </Option.Avatar>
+        <Option.Content>
+          Lorem Ipsum Lorem Lorem Ipsum Lorem Lorem Ipsum Lorem Lorem Ipsum
+          Lorem
+        </Option.Content>
       </Option>
       <Option>
         <Option.Avatar>
           <Avatar url={exampleAvatar} size='x28' />
-        </Option.Avatar>        
-        <Option.Content>Lorem Ipsum Lorem <Option.Description>Lorem Lorem Ipsum Lorem Lorem Ipsum Lorem</Option.Description></Option.Content>
+        </Option.Avatar>
+        <Option.Content>
+          Lorem Ipsum Lorem{' '}
+          <Option.Description>
+            Lorem Lorem Ipsum Lorem Lorem Ipsum Lorem
+          </Option.Description>
+        </Option.Content>
       </Option>
     </Box>
   </Story>
@@ -60,16 +76,23 @@ The generic <LinkTo kind='Misc/Options/Option' story='Default'>`Option`</LinkTo>
 <Canvas>
   <Story name='withPresence'>
     <Box position='relative' maxWidth={250}>
-      <Option>
-        <Option.Column><StatusBullet /></Option.Column>
+      <Option onClick={action('click')}>
+        <Option.Column>
+          <StatusBullet />
+        </Option.Column>
         <Option.Content>Lorem Ipsum Lorem</Option.Content>
       </Option>
       <Option>
         <Option.Avatar>
           <Avatar url={exampleAvatar} size='x28' />
         </Option.Avatar>
-        <Option.Column><StatusBullet /></Option.Column>
-        <Option.Content>Lorem Ipsum Lorem Lorem Ipsum Lorem Lorem Ipsum Lorem Lorem Ipsum Lorem</Option.Content>
+        <Option.Column>
+          <StatusBullet />
+        </Option.Column>
+        <Option.Content>
+          Lorem Ipsum Lorem Lorem Ipsum Lorem Lorem Ipsum Lorem Lorem Ipsum
+          Lorem
+        </Option.Content>
       </Option>
     </Box>
   </Story>
@@ -80,17 +103,26 @@ The generic <LinkTo kind='Misc/Options/Option' story='Default'>`Option`</LinkTo>
 <Canvas>
   <Story name='withMenu'>
     <Box position='relative' maxWidth={250}>
-      <Option>
+      <Option onClick={action('click')}>
         <Option.Content>Lorem Ipsum Lorem</Option.Content>
-        <Option.Menu><Menu options={menuOptions} /></Option.Menu>
+        <Option.Menu>
+          <Menu options={menuOptions} />
+        </Option.Menu>
       </Option>
       <Option>
         <Option.Avatar>
           <Avatar url={exampleAvatar} size='x28' />
         </Option.Avatar>
-        <Option.Column><StatusBullet /></Option.Column>
-        <Option.Content>Lorem Ipsum Lorem Lorem Ipsum Lorem Lorem Ipsum Lorem Lorem Ipsum Lorem</Option.Content>
-        <Option.Menu><Menu options={menuOptions} /></Option.Menu>
+        <Option.Column>
+          <StatusBullet />
+        </Option.Column>
+        <Option.Content>
+          Lorem Ipsum Lorem Lorem Ipsum Lorem Lorem Ipsum Lorem Lorem Ipsum
+          Lorem
+        </Option.Content>
+        <Option.Menu>
+          <Menu options={menuOptions} />
+        </Option.Menu>
       </Option>
     </Box>
   </Story>
@@ -101,19 +133,28 @@ The generic <LinkTo kind='Misc/Options/Option' story='Default'>`Option`</LinkTo>
 <Canvas>
   <Story name='withIcon'>
     <Box position='relative' maxWidth={250}>
-      <Option>
-        <Option.Icon name='bell'/>
+      <Option onClick={action('click')}>
+        <Option.Icon name='bell' />
         <Option.Content>Lorem Ipsum Lorem</Option.Content>
-        <Option.Menu><Menu options={menuOptions} /></Option.Menu>
+        <Option.Menu>
+          <Menu options={menuOptions} />
+        </Option.Menu>
       </Option>
       <Option>
         <Option.Avatar>
           <Avatar url={exampleAvatar} size='x28' />
         </Option.Avatar>
-        <Option.Icon name='bell'/>
-        <Option.Column><StatusBullet /></Option.Column>
-        <Option.Content>Lorem Ipsum Lorem Lorem Ipsum Lorem Lorem Ipsum Lorem Lorem Ipsum Lorem</Option.Content>
-        <Option.Menu><Menu options={menuOptions} /></Option.Menu>
+        <Option.Icon name='bell' />
+        <Option.Column>
+          <StatusBullet />
+        </Option.Column>
+        <Option.Content>
+          Lorem Ipsum Lorem Lorem Ipsum Lorem Lorem Ipsum Lorem Lorem Ipsum
+          Lorem
+        </Option.Content>
+        <Option.Menu>
+          <Menu options={menuOptions} />
+        </Option.Menu>
       </Option>
     </Box>
   </Story>
@@ -123,17 +164,25 @@ The generic <LinkTo kind='Misc/Options/Option' story='Default'>`Option`</LinkTo>
 
 <Canvas>
   <Story name='asUserItem'>
-    <Box maxWidth={330}>
-      <Option>
+    <Box position='relative' maxWidth={330}>
+      <Option onClick={action('click')}>
         <Option.Avatar>
           <Avatar url={exampleAvatar} size='x28' />
-        </Option.Avatar> 
-        
-        <Option.Column><StatusBullet status='online' /></Option.Column>
+        </Option.Avatar>
+        <Option.Column>
+          <StatusBullet />
+        </Option.Column>
         <Option.Content>
-          <Box withTruncatedText fontScale='p3'>carla.culhane <Box is='span' color='neutral-500'>(carla hune)</Box></Box>
+          <Box withTruncatedText fontScale='p3'>
+            carla.culhane{' '}
+            <Box is='span' color='neutral-500'>
+              (carla hune)
+            </Box>
+          </Box>
         </Option.Content>
-        <Option.Menu><Menu options={menuOptions} /></Option.Menu>
+        <Option.Menu>
+          <Menu options={menuOptions} />
+        </Option.Menu>
       </Option>
     </Box>
   </Story>
@@ -144,11 +193,9 @@ The generic <LinkTo kind='Misc/Options/Option' story='Default'>`Option`</LinkTo>
 <Canvas>
   <Story name='Skeleton'>
     <Box position='relative' maxWidth={330}>
-      <Option.Skeleton /> 
+      <Option.Skeleton />
     </Box>
   </Story>
 </Canvas>
 
 <ArgsTable of={Option} />
-
-
