--- packages/fuselage/src/components/OptionsPaginated/Option/Option.stories.mdx
@@ -1,154 +1,0 @@
-import { Meta, Canvas, ArgsTable, Story } from '@storybook/addon-docs/blocks';
-import LinkTo from '@storybook/addon-links/react';
-
-import { Option } from '.';
-import { Box, Avatar, Menu, StatusBullet } from '../..';
-import { exampleAvatar, menuOptions } from '../../../../.storybook/helpers.js';
-
-<Meta title='Misc/Options/Option' parameters={{ jest: ['Option/spec'] }} />
-
-# Option
-
-The generic <LinkTo kind='Misc/Options/Option' story='Default'>`Option`</LinkTo> item of options. Can be freely used or inside the <LinkTo kind='Misc/Options/Options' story='Default'>`Options`</LinkTo> as well.
-
-<Canvas>
-  <Story name='Default'>
-    <Box position='relative' maxWidth={250}>
-      <Option>
-        <Option.Content>Lorem Ipsum Lorem</Option.Content>
-      </Option>
-      <Option>
-        <Option.Content>Lorem Ipsum Lorem Lorem Ipsum Lorem Lorem Ipsum Lorem Lorem Ipsum Lorem</Option.Content>
-      </Option>
-      <Option>
-        <Option.Content>Lorem Ipsum Lorem <Option.Description>Lorem Lorem Ipsum Lorem Lorem Ipsum Lorem</Option.Description></Option.Content>
-      </Option>
-    </Box>
-  </Story>
-</Canvas>
-
-
-## With avatar
-
-<Canvas>
-  <Story name='withAvatar'>
-    <Box position='relative' maxWidth={250}>
-      <Option>
-        <Option.Avatar>
-          <Avatar url={exampleAvatar} size='x28' />
-        </Option.Avatar>
-        <Option.Content>Lorem Ipsum Lorem</Option.Content>
-      </Option>
-      <Option>
-        <Option.Avatar>
-          <Avatar url={exampleAvatar} size='x28' />
-        </Option.Avatar>        
-        <Option.Content>Lorem Ipsum Lorem Lorem Ipsum Lorem Lorem Ipsum Lorem Lorem Ipsum Lorem</Option.Content>
-      </Option>
-      <Option>
-        <Option.Avatar>
-          <Avatar url={exampleAvatar} size='x28' />
-        </Option.Avatar>        
-        <Option.Content>Lorem Ipsum Lorem <Option.Description>Lorem Lorem Ipsum Lorem Lorem Ipsum Lorem</Option.Description></Option.Content>
-      </Option>
-    </Box>
-  </Story>
-</Canvas>
-
-## With presence
-
-<Canvas>
-  <Story name='withPresence'>
-    <Box position='relative' maxWidth={250}>
-      <Option>
-        <Option.Column><StatusBullet /></Option.Column>
-        <Option.Content>Lorem Ipsum Lorem</Option.Content>
-      </Option>
-      <Option>
-        <Option.Avatar>
-          <Avatar url={exampleAvatar} size='x28' />
-        </Option.Avatar>
-        <Option.Column><StatusBullet /></Option.Column>
-        <Option.Content>Lorem Ipsum Lorem Lorem Ipsum Lorem Lorem Ipsum Lorem Lorem Ipsum Lorem</Option.Content>
-      </Option>
-    </Box>
-  </Story>
-</Canvas>
-
-## With menu
-
-<Canvas>
-  <Story name='withMenu'>
-    <Box position='relative' maxWidth={250}>
-      <Option>
-        <Option.Content>Lorem Ipsum Lorem</Option.Content>
-        <Option.Menu><Menu options={menuOptions} /></Option.Menu>
-      </Option>
-      <Option>
-        <Option.Avatar>
-          <Avatar url={exampleAvatar} size='x28' />
-        </Option.Avatar>
-        <Option.Column><StatusBullet /></Option.Column>
-        <Option.Content>Lorem Ipsum Lorem Lorem Ipsum Lorem Lorem Ipsum Lorem Lorem Ipsum Lorem</Option.Content>
-        <Option.Menu><Menu options={menuOptions} /></Option.Menu>
-      </Option>
-    </Box>
-  </Story>
-</Canvas>
-
-## With Icon
-
-<Canvas>
-  <Story name='withIcon'>
-    <Box position='relative' maxWidth={250}>
-      <Option>
-        <Option.Icon name='bell'/>
-        <Option.Content>Lorem Ipsum Lorem</Option.Content>
-        <Option.Menu><Menu options={menuOptions} /></Option.Menu>
-      </Option>
-      <Option>
-        <Option.Avatar>
-          <Avatar url={exampleAvatar} size='x28' />
-        </Option.Avatar>
-        <Option.Icon name='bell'/>
-        <Option.Column><StatusBullet /></Option.Column>
-        <Option.Content>Lorem Ipsum Lorem Lorem Ipsum Lorem Lorem Ipsum Lorem Lorem Ipsum Lorem</Option.Content>
-        <Option.Menu><Menu options={menuOptions} /></Option.Menu>
-      </Option>
-    </Box>
-  </Story>
-</Canvas>
-
-## As UserItem
-
-<Canvas>
-  <Story name='asUserItem'>
-    <Box maxWidth={330}>
-      <Option>
-        <Option.Avatar>
-          <Avatar url={exampleAvatar} size='x28' />
-        </Option.Avatar> 
-        
-        <Option.Column><StatusBullet status='online' /></Option.Column>
-        <Option.Content>
-          <Box withTruncatedText fontScale='p3'>carla.culhane <Box is='span' color='neutral-500'>(carla hune)</Box></Box>
-        </Option.Content>
-        <Option.Menu><Menu options={menuOptions} /></Option.Menu>
-      </Option>
-    </Box>
-  </Story>
-</Canvas>
-
-## As Skeleton
-
-<Canvas>
-  <Story name='Skeleton'>
-    <Box position='relative' maxWidth={330}>
-      <Option.Skeleton /> 
-    </Box>
-  </Story>
-</Canvas>
-
-<ArgsTable of={Option} />
-
-
