+++ packages/fuselage/src/components/OptionsPaginated/OptionsPaginated.stories.mdx
@@ -1,8 +1,8 @@
-import { Meta, Canvas, Story } from '@storybook/addon-docs/blocks';
+import { Meta, Canvas, Story } from '@storybook/addon-docs';
 
 import { OptionsPaginated, CheckOption } from '.';
 import { Box } from '..';
-import { Option } from './Option';
+import { Option } from '../Options/Option';
 
 export const options = [
   { value: 1, label: 'a teste 1', selected: true },
